+++ .changeset/tasty-loops-lose.md
@@ -1,0 +1,6 @@
+---
+"@verdaccio/middleware": patch
+"@verdaccio/loaders": patch
+---
+
+chore: rename files with underscore
