+++ packages/loaders/test/plugin-loader-async.spec.ts
